Added listing and removing plugins via cli
