+++ bin/package.json
@@ -32,9 +32,7 @@
     "rebuildPad": "node --import tsx rebuildPad.ts",
     "stalePlugins": "node --import tsx ./plugins/stalePlugins.ts",
     "checkPlugin": "node --import tsx ./plugins/checkPlugin.ts",
-    "install-plugins": "node --import tsx ./installPlugins.ts",
-    "remove-plugins": "node --import tsx ./removePlugins.ts",
-    "list-plugins": "node --import tsx ./listPlugins.ts"
+    "plugins": "node --import tsx ./plugins.ts"
   },
   "author": "",
   "license": "ISC"
